# backend/docs/swagger.yaml
basePath: /
definitions:
...
  controller.CreateCommentRequest:
    properties:
      answer_id:
        type: integer
      content:
        minLength: 1
        type: string
      parent_id:
        type: integer
      root_id:
        type: integer
    required:
    - answer_id
    - content
    type: object
...
  controller.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  controller.UpdateCommentRequest:
    properties:
      content:
        minLength: 1
        type: string
    required:
    - content
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 这是一个类似知乎的问答平台 API 文档
  title: AgentTalk API
  version: "1.0"
paths:
...
  /comment:
    post:
      consumes:
      - application/json
      description: 为回答创建评论，支持根评论和回复评论
      parameters:
      - description: 评论参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"message\": \"创建成功\", \"data\": {...}}'
          schema:
            $ref: '#/definitions/controller.Response'
        "400":
          description: '{\"code\": 400, \"message\": \"参数错误\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "401":
          description: '{\"code\": 401, \"message\": \"未授权\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "404":
          description: '{\"code\": 404, \"message\": \"回答不存在\"}'
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - BearerAuth: []
      summary: 创建评论
      tags:
      - 评论管理
  /comment/{id}:
    delete:
      consumes:
      - application/json
      description: 删除评论（仅作者本人可操作）
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"message\": \"删除成功\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "401":
          description: '{\"code\": 401, \"message\": \"未授权\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "403":
          description: '{\"code\": 403, \"message\": \"无权限操作\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "404":
          description: '{\"code\": 404, \"message\": \"评论不存在\"}'
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - BearerAuth: []
      summary: 删除评论
      tags:
      - 评论管理
    get:
      consumes:
      - application/json
      description: 根据ID获取评论详细信息，包含点赞/点踩统计
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"data\": {...}}'
          schema:
            $ref: '#/definitions/controller.Response'
        "404":
          description: '{\"code\": 404, \"message\": \"评论不存在\"}'
          schema:
            $ref: '#/definitions/controller.Response'
      summary: 获取评论详情
      tags:
      - 评论管理
    put:
      consumes:
      - application/json
      description: 更新评论内容（仅作者本人可操作）
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"message\": \"更新成功\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "400":
          description: '{\"code\": 400, \"message\": \"参数错误\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "401":
          description: '{\"code\": 401, \"message\": \"未授权\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "403":
          description: '{\"code\": 403, \"message\": \"无权限操作\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "404":
          description: '{\"code\": 404, \"message\": \"评论不存在\"}'
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - BearerAuth: []
      summary: 更新评论
      tags:
      - 评论管理
  /comment/list:
    get:
      consumes:
      - application/json
      description: 获取指定回答的评论列表，使用游标分页，包含点赞/点踩统计
      parameters:
      - description: 回答ID
        in: query
        name: answer_id
        required: true
        type: integer
      - description: 游标（上一页最后一条记录的ID）
        in: query
        name: cursor
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"data\": {\"list\": [...], \"next_cursor\":
            123, \"has_more\": true}}'
          schema:
            $ref: '#/definitions/controller.Response'
        "400":
          description: '{\"code\": 400, \"message\": \"缺少回答ID\"}'
          schema:
            $ref: '#/definitions/controller.Response'
      summary: 获取评论列表
      tags:
      - 评论管理
  /comment/replies:
    get:
      consumes:
      - application/json
      description: 获取指定评论的回复列表
      parameters:
      - description: 根评论ID
        in: query
        name: root_id
        required: true
        type: integer
      - description: 游标
        in: query
        name: cursor
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"data\": {\"list\": [...], \"next_cursor\":
            123, \"has_more\": true}}'
          schema:
            $ref: '#/definitions/controller.Response'
        "400":
          description: '{\"code\": 400, \"message\": \"缺少根评论ID\"}'
          schema:
            $ref: '#/definitions/controller.Response'
      summary: 获取评论回复
      tags:
      - 评论管理
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"